web/secret_web/deploy/app/requirements.txt
asyncpg
fastapi
ormar
psycopg2-binary
pydantic-settings
uvicorn
starlette
pydantic
databases
SQLAlchemy
jinja2
pyjwt
python-multipart